ingestion/requirements.txt
psycopg2-binary
langchain-openai
langchain-core
lxml
numpy